feat(bevy_components):  * fixed issue with missing types lists being cleared when reloading registry  * refactored/moved around ui elements  * minor cleanups
